Move pipeline publisher descriptor class inside publisher class
