+++ src/main/java/hudson/plugins/plot/pipeline/PlotPublisher.java
@@ -4,17 +4,24 @@
  */
 package hudson.plugins.plot.pipeline;
 
+import hudson.Extension;
 import hudson.FilePath;
 import hudson.Launcher;
 import hudson.Util;
+import hudson.matrix.MatrixProject;
+import hudson.model.AbstractProject;
+import hudson.model.Job;
 import hudson.model.Run;
 import hudson.model.TaskListener;
 import hudson.plugins.plot.AbstractPlotPublisher;
+import hudson.plugins.plot.Messages;
 import hudson.plugins.plot.Plot;
+import hudson.plugins.plot.SeriesFactory;
 import hudson.tasks.BuildStepDescriptor;
 import hudson.tasks.BuildStepMonitor;
 import hudson.tasks.Publisher;
 import hudson.tasks.Recorder;
+import hudson.util.FormValidation;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -24,8 +31,12 @@ import java.util.Map;
 import java.util.logging.Logger;
 import javax.annotation.Nonnull;
 import jenkins.tasks.SimpleBuildStep;
+import net.sf.json.JSONObject;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang.StringUtils;
+import org.kohsuke.stapler.AncestorInPath;
+import org.kohsuke.stapler.QueryParameter;
+import org.kohsuke.stapler.StaplerRequest;
 
 /**
  * Records the plotpipeline data for builds.
@@ -150,14 +161,6 @@ public class PlotPublisher extends Recorder implements SimpleBuildStep {
     }
 
     /**
-     * Called by Jenkins.
-     */
-    @Override
-    public BuildStepDescriptor<Publisher> getDescriptor() {
-        return DESCRIPTOR;
-    }
-
-    /**
      * Called by Jenkins when a build is finishing.
      */
     @Override
@@ -179,5 +182,72 @@ public class PlotPublisher extends Recorder implements SimpleBuildStep {
         return BuildStepMonitor.BUILD;
     }
 
+    /**
+     * Called by Jenkins.
+     */
+    @Override
+    public BuildStepDescriptor<Publisher> getDescriptor() {
+        return DESCRIPTOR;
+    }
+
+    @Extension
     public static final PlotDescriptor DESCRIPTOR = new PlotDescriptor();
+
+    /**
+     * The Descriptor for the plotpipeline configuration Extension
+     *
+     * @author Nigel Daley
+     * @author Thomas Fox
+     *
+     */
+    public static class PlotDescriptor extends BuildStepDescriptor<Publisher> {
+
+        public PlotDescriptor() {
+            super(PlotPublisher.class);
+        }
+
+        public String getDisplayName() {
+            return Messages.Plot_Pipeline_Publisher_DisplayName();
+        }
+
+        @Override
+        public boolean isApplicable(Class<? extends AbstractProject> jobType) {
+            return AbstractProject.class.isAssignableFrom(jobType)
+                    && !MatrixProject.class.isAssignableFrom(jobType);
+        }
+
+        /**
+         * Called when the user saves the project configuration.
+         */
+        @Override
+        public Publisher newInstance(StaplerRequest req, JSONObject formData)
+                throws FormException {
+            PlotPublisher publisher = new PlotPublisher();
+            for (Object data : SeriesFactory.getArray(formData.get("plots"))) {
+                publisher.addPlot(bindPlot((JSONObject) data, req));
+            }
+            return publisher;
+        }
+
+        private static Plot bindPlot(JSONObject data, StaplerRequest req) {
+            Plot p = req.bindJSON(Plot.class, data);
+            p.series = SeriesFactory.createSeriesList(data.get("series"), req);
+            return p;
+        }
+
+        /**
+         * Checks if the series file is valid.
+         * Called from "config.jelly".
+         */
+        public FormValidation doCheckSeriesFile(
+                @AncestorInPath Job<?, ?> project,
+                @QueryParameter String value) throws IOException {
+            FilePath fp = new FilePath(new FilePath(project.getRootDir()), "workspace" );
+            // Check if workspace folder is missing form root directory
+            if (fp.validateFileMask(value) == null) {
+                return new FilePath(project.getRootDir()).validateFileMask(value);
+            }
+            return fp.validateFileMask(value);
+        }
+    }
 }
