+++ src/main/resources/hudson/plugins/plot/PropertiesSeries/config.jelly
@@ -4,7 +4,7 @@
     <table width="100%" bgcolor="#EEEEEE">
     <f:entry title="${%Data series file}" help="/plugin/plot/help-series.html">
         <f:textbox name="file" value="${series.file}"
-                   checkUrl="'descriptorByName/PlotPublisher/checkSeriesFile?value='+toValue(this)" />
+                   checkUrl="'descriptorByName/hudson.plugins.plot.pipeline.PlotPublisher/checkSeriesFile?value='+toValue(this)" />
     </f:entry>
         <f:nested>
             <f:entry title="${%Data series legend label}" help="/plugin/plot/help-legend.html">
